поправил переполнение при сложении двух чисел

+++ 01_week/tasks/addition/addition.cpp
@@ -3,6 +3,6 @@
 
 
 int64_t Addition(int a, int b) {
-    int64_t c=a+b;
+    int64_t c=static_cast<int64_t>(a)+static_cast<int64_t>(b);
     return c;
 }
